AOC-2015: Rebuild dec_07 to work for real input
The previous solution was only working for the test input,
which had all variables coming in order. No variable
was used before set, which is not the case in the real
input.

Storing all input in a list is now working, next stage is
to go through the list and and pop nodes were alla signals
are known and keep doing this until the list is empty.